+++ components/client/Camera.tsx
