+++ components/client/LoginForm.tsx
@@ -5,7 +5,7 @@ import { Pressable, View, Text, StyleSheet } from 'react-native';
 import { useRouter } from 'expo-router';
 
 import { useSession, signIn } from '@/lib/auth-client';
-import { LoadingSpinner } from '@/components/LoadingSpinner';
+import { LoadingSpinner } from '@/components/client/LoadingSpinner';
 
 export const LoginForm = () => {
   const [isSubmitting, setIsSubmitting] = useState(false);
