+++ components/client/PhotoSelector.tsx
