+++ components/client/SplashScreenController.tsx
