+++ components/ui/gluestack-ui-provider/index.web.tsx
@@ -31,7 +31,7 @@ export function GluestackUIProvider({
     cssVariablesWithMode +=
       configKey === 'dark' ? `\n .dark {\n ` : `\n:root {\n`;
     const cssVariables = Object.keys(
-      config[configKey as keyof typeof config]
+      config[configKey as keyof typeof config],
     ).reduce((acc: string, curr: string) => {
       acc += `${curr}:${config[configKey as keyof typeof config][curr]}; `;
       return acc;
